backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
pydantic==2.9.2
python-multipart==0.0.9
openpyxl==3.1.5
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
